types and styles
