+++ src/frontend/src/views/Data/components/SamplesView/components/SampleTableModalManager/components/IconButton/style.ts
